pyproject.toml
[project]
name = "job-agent"
version = "0.1.0"
description = "Prompt-driven autonomous job discovery and application assistant."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "beautifulsoup4>=4.12.3",
  "httpx>=0.27.0",
  "playwright>=1.52.0",
  "pydantic>=2.9.0",
  "pyyaml>=6.0.2",
  "rich>=13.9.0",
  "tenacity>=9.0.0",
  "typer>=0.16.0",
]

[project.optional-dependencies]
llm = [
  "openai>=1.75.0",
]

[project.scripts]
job-agent = "job_agent.cli:app"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
